renamed files to reflect new class names
