+++ CMakeLists.txt
@@ -36,13 +36,9 @@ catkin_package(
 
 include_directories(common/include ${Boost_INCLUDE_DIRS} ${catkin_INCLUDE_DIRS})
 
-add_library(SerialIO common/src/SerialIO.cpp)
-add_library(${PROJECT_NAME} common/src/RelayBoard_v2.cpp common/src/StrUtil.cpp common/src/neo_relayboard_v2_lib.cpp)
+add_library(${PROJECT_NAME} common/src/RelayBoardClient.cpp common/src/NeoRelayBoardNode.cpp)
 
 add_dependencies(${PROJECT_NAME} ${PROJECT_NAME}_gencpp)
-add_dependencies(SerialIO ${PROJECT_NAME}_gencpp)
-
-target_link_libraries(${PROJECT_NAME} SerialIO)
 
 add_executable(neo_relayboard_v2_node src/neo_relayboard_v2_node.cpp)
 target_link_libraries(neo_relayboard_v2_node ${PROJECT_NAME})
